create ElementModificatorInterface.h
